First pass at celery-ifying test runner
