+++ requirements.txt
@@ -7,8 +7,12 @@ MarkupSafe==0.23
 PyVirtualDisplay==0.2
 PyYAML==3.11
 Werkzeug==0.11.9
+amqp==1.4.9
+anyjson==0.3.3
 astroid==1.4.5
 behave==1.2.5
+billiard==3.3.0.23
+celery==3.1.23
 colorama==0.3.7
 cookies==2.2.1
 coverage==4.1
@@ -17,6 +21,7 @@ flask-log==0.1.0
 httpretty==0.8.14
 isodate==0.5.4
 itsdangerous==0.24
+kombu==3.0.35
 lazy-object-proxy==1.2.2
 parse==1.6.6
 parse-type==0.3.4
@@ -27,6 +32,8 @@ pytest==2.9.1
 pytest-cov==2.2.1
 python-engineio==0.9.1
 python-socketio==1.3
+pytz==2016.4
+redis==2.10.5
 requests==2.9.1
 selenium==2.53.2
 six==1.10.0
