+++ runserver.py
@@ -1,3 +1,4 @@
-from testsuite import app, socketio
+from testsuite import main
 
-socketio.run(app, host='0.0.0.0', port=5000)
+app, socketio = main()
+socketio.run(app, host='0.0.0.0', port=5000, debug=True, use_reloader=True)
