+++ runserver.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
 
 pip install -r requirements.txt
-python -m runserver
+supervisorctl start after:*
