+++ supervisord.conf
@@ -1,0 +1,46 @@
+[unix_http_server]
+file=/tmp/supervisor.sock                       ; path to your socket file
+
+[supervisord]
+logfile=/var/log/supervisor/supervisord.log     ; supervisord log file
+logfile_maxbytes=50MB                           ; maximum size of logfile before rotation
+logfile_backups=10                              ; number of backed up logfiles
+loglevel=error                                  ; info, debug, warn, trace
+pidfile=/var/run/supervisord.pid                ; pidfile location
+nodaemon=true                                   ; run supervisord as a daemon
+minfds=1024                                     ; number of startup file descriptors
+minprocs=200                                    ; number of process descriptors
+user=root                                       ; default user
+childlogdir=/var/log/supervisor/                ; where child log files will live
+
+[rpcinterface:supervisor]
+supervisor.rpcinterface_factory = supervisor.rpcinterface:make_main_rpcinterface
+
+[supervisorctl]
+serverurl=unix:///tmp/supervisor.sock         ; use a unix:// URL  for a unix socket
+
+; Program "before" will run once, doing any init tasks and then finish with
+; a call to `supervisorctl start after:*`. This will start any other tasks
+; that need to wait until the environment is ready.
+
+[group:after]
+programs=celery,flask
+
+[program:before]
+command=./runserver.sh
+directory=/demo-test-suite
+autorestart=false
+
+[program:redis]
+command=/usr/bin/redis-server
+
+[program:celery]
+command=/usr/bin/env celery -A testsuite.tasks worker --config=testsuite.celeryconfig
+directory=/demo-test-suite
+environment=C_FORCE_ROOT="yes"
+autostart=false
+
+[program:flask]
+command=/usr/bin/env python -m "runserver"
+directory=/demo-test-suite
+autostart=false
